:recycle: :octocat: solve race condition on lambda deployment
Since the direktiv's deployment happens in parallel with the
slack notification, it may be terminated before everything else
since slack bot checks the current status of the workflow and
sets a final status. If that happens before the completion of
the release of the direktiv's code execution, it would not
execute the HTTP call to the refresh endpoint.