:octocat: :bug: release version with current releases
The calls fail with the parsing of the text from the api response.
For this reason, we can just write to a file, remove the body
property and then parse it correctly with jq since the call
productes a json with a line for each property!

Using https://stackoverflow.com/questions/61576308/how-do-i-have-quotes-inside-of-a-echo/61588193#61588193 to fix