:octocat: :bug: make sure SHA is for commit in branch
The pull request SHA is different from the push event SHA because they are associated with different Git refs. In a push event, the SHA corresponds to the SHA of the commit that was pushed to the Git ref. In a pull request event, the SHA corresponds to the SHA of the head commit of the pull request branch.

To get the SHA of the commit associated with the Git ref that was pushed to, you can use the `${{ github.sha }}` syntax in your workflow, like this:

```
- name: Get commit SHA
  run: echo ${{ github.sha }}
```

This will output the SHA of the commit that was pushed to the Git ref that triggered the workflow.

To get the SHA of the head commit of the pull request branch, you can use the `pull_request` event trigger in your workflow, like this:

```
on:
  pull_request:
    types: [opened, synchronize, closed]
jobs:
  my-job:
    runs-on: ubuntu-latest
    steps:
      - name: Get head SHA
        run: echo ${{ github.event.pull_request.head.sha }}
```

This will output the SHA of the head commit of the pull request branch that triggered the workflow.

If you want to use the same SHA in both types of events, you can modify your Git workflow to use a branch ref that is used both for pull requests and for direct pushes. This way, the same commit will be associated with both the push event and the pull request event, and you can use `${{ github.sha }}` to get the SHA in both types of events.

For example, you can use the `main` branch as the ref for all your Git operations, and set up your pull requests to merge changes into `main`. Then, you can use the following workflow to get the SHA of the latest commit in `main`, regardless of whether the changes were pushed directly to `main` or submitted as a pull request:

```
on:
  push:
    branches:
      - main
  pull_request:
    types: [opened, synchronize, closed]
jobs:
  my-job:
    runs-on: ubuntu-latest
    steps:
      - name: Get commit SHA
        run: echo ${{ github.sha }}
```

In this example, the workflow runs on both `push` events to the `main` branch and `pull_request` events, and `${{ github.sha }}` is used to get the SHA of the latest commit in the `main` branch in both types of events.