reference the current branch for code testing
currently, the docker compose check only checks out the main branch for testing. This makes any change on the compose not runnable until it bubbles up to the main branch. This ref will allow changes to the docker compose test to be ran within the PR for the current branch