:lock: :bug: don't expose kubeconfig file with base64
Using bash here-string to hide the command execution of base64 -d.
This is because the workflow would print the file location being
in HOME/.kube/config. Instead, we create a random file, add it to
the github mask in addition to the value of kubeconfig itself. This
won't leak the information in the github actions workflow logs.