:octocar: :bug: Fix action run logs download with exponential backoff
The download of the workflow run logs is eventual consistent and, for this reason, it's not always ready for download. For this reason, the client needs to use the exponential backoff strategy to properly download the zip file when it's available, but slowly increasing the time to wait for the resource.